test(RM): test interface deletion
Adds test to check for interface deletion. This should happen only if
the major version of such interface is >= 0. Interface deletion means
that the interface does not show up when listing all available
interfaces

Adds test to check for interface update. An update can only ahppen to a
minor version to another, editing only some parameters of the interface.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>